+++ ElvUI_LuckyoneUI/Core/Setup.lua
@@ -171,12 +171,18 @@ function Private:Setup_PrivateDB()
 	E.private.nameplates.enable = false
 	E.private.skins.parchmentRemoverEnable = true
 
-	if E.Retail and E:IsAddOnEnabled('ElvUI_SLE') then
-		Private:Setup_Private_ShadowAndLight()
+	if E.Retail then
+		if E:IsAddOnEnabled('ElvUI_SLE') then
+			Private:Setup_Private_ShadowAndLight()
+		end
+		if E:IsAddOnEnabled('ElvUI_WindTools') then
+			Private:Setup_Private_WindTools()
+		end
 	end
 
-	if E.Retail and E:IsAddOnEnabled('ElvUI_WindTools') then
-		Private:Setup_Private_WindTools()
+	if E.Classic then
+		-- This will make sure Shaman is blue instead of pink
+		E.private.general.classColors = true
 	end
 
 	if E.global.L1UI.dev then