+++ ElvUI_LuckyoneUI/Profiles/OmniCD.lua
@@ -11,7 +11,10 @@ local _, Private = ...
 -- ElvUI modules
 local E, L = unpack(ElvUI)
 
--- OmniCD profile
+--[[
+	OmniCD profile
+	LC: 29/12/2024
+]]
 function Private:Setup_OmniCD(layout, installer)
 	if not E:IsAddOnEnabled('OmniCD') and E.Retail then Private:Print('OmniCD ' .. L["is not installed or enabled."]) return end