+++ CHANGELOG.md
@@ -1,6 +1,18 @@
-### Version 3.13 [In Development]
+### Version 3.14 [In Development]
+**All WoW Versions:**  
+ElvUI 13.84+ compatibility  
+**Retail:**  
+Updates for Shadow & Light profile (AFK Mode, Armory Mode)  
+Updates for WindTools profile (Disabled Map EventTracker)  
+___
+### Version 3.13 [29-01-2025]
+**All WoW Versions:**  
+ElvUI 13.82+ compatibility  
 **Retail:**  
 Minor pixel-perfect edit for OmniCD icons  
+**Classic:**  
+Shaman default color set to blue instead of pink  
+Updated TOC for 1.15.6  
 **Misc:**  
 Updated LICENSE  
 ___